Add switchdev-online service to hold off on network init
The addresses a race condition between the switchdev driver, udev
and ifupdown2. In some cases, ifupdown2 attempts to configure
switchdev interfaces (1) before they are ready, or (2) before any
udev rules have had a chance to rename them per udev configuration.
When this happens, not all interfaces are configured correctly.

The solution is to

1. create a dummy systemd service for each switchdev interface
   (platform specific) that depends on the well-known systemd
   device path for that netdev (currently
   'sys-subsystem-net-devices-FOO.device'.
2. Make the system-wide network-pre.service depend on *all*
   switchdev devices coming online
3. make ifupdown2 (via networking.service) depend on network-pre.service

See
https://www.freedesktop.org/wiki/Software/systemd/NetworkTarget/

Signed-off-by: Carl Roth <rothcar@amazon.com>