buster: drop broken netdev udev rule
The udev rule 60-net.rules tries to call '/sbin/initnetdev', but that
file is found at a different location, at '/usr/bin/initnetdev', so the
call always fails.

Since the only purpose of initnetdev is to rename network devices based
on the platform configuration, the initial management port renaming is
already done by the initrd, and for any subsequent port devices we have
an extra rule, drop the udev rule instead of trying to fix it.

Silences errors at boot like:

> Sep 15 09:18:55 localhost systemd-udevd[573]: Process '/sbin/initnetdev ma1 add' failed with exit code 1.
> Sep 15 09:18:56 localhost systemd-udevd[553]: Process '/sbin/initnetdev eth1 add' failed with exit code 1.
> Sep 15 09:18:56 localhost systemd-udevd[586]: Process '/sbin/initnetdev eth11 add' failed with exit code 1.
> Sep 15 09:18:56 localhost systemd-udevd[594]: Process '/sbin/initnetdev eth33 add' failed with exit code 1.
> Sep 15 09:18:56 localhost systemd-udevd[596]: Process '/sbin/initnetdev eth19 add' failed with exit code 1.
> ...

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>