kernel/5.15-lts: import dts(i) files from kernel tree
The DentOS own DTS files are in no shape to be upstreamed, and there
likely is no one willing to spend the effort to do so, so there is no
good reason to carry them as commits for the kernel source.

Instead provide them within the build system, as they are independent
from the kernel.

This also makes it easier to provide new device support without the need
to update the kernel, or atomic changes of device drivers and the
associated device trees.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>